resolve CR comments
