+++ rlib-logger-impl/build.gradle
@@ -6,4 +6,5 @@ plugins {
 dependencies {
   api projects.rlibCommon
   api projects.rlibCollections
+  testImplementation testFixtures(projects.rlibCommon)
 }
